+++ project-keeper/src/main/java/com/exasol/projectkeeper/github/FileContentProvider.java
@@ -1,4 +1,4 @@
-package com.exasol.projectkeeper.githuboutput;
+package com.exasol.projectkeeper.github;
 
 import java.io.*;
 import java.nio.charset.StandardCharsets;
