+++ project-keeper/src/main/java/com/exasol/projectkeeper/github/OutputPublisherFactory.java
@@ -1,4 +1,4 @@
-package com.exasol.projectkeeper.githuboutput;
+package com.exasol.projectkeeper.github;
 
 import java.nio.file.Path;
 import java.util.Map;
