+++ project-keeper/src/test/java/com/exasol/projectkeeper/github/KeyValuePairFormatterTest.java
@@ -1,4 +1,4 @@
-package com.exasol.projectkeeper.githuboutput;
+package com.exasol.projectkeeper.github;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.equalTo;
@@ -7,6 +7,8 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.ValueSource;
+
+import com.exasol.projectkeeper.github.KeyValuePairFormatter;
 
 // [utest->dsn~verify-modes.output-parameters~1]
 class KeyValuePairFormatterTest {
